Use guest user if not signed in.
The application controller will look for a current_user via devise
and assign it to @current_user. If no session is found, it will
assign @current_user to the guest account that is created in the seeds.

An anonymous application controller class is added in the tests to
test this before_action method.