Create identicons when adding user content
When a guest adds a new discussion or comment then
check to see if the user has a guest_id cookie. If
not then create one. 

Discussions and DiscussionComments get a guest_id
column that get populated via a hidden field in the
existing forms.

Identicons are created for each unique guest cookie
and uploaded to amazon.

When a guest registers, a sidekiq worker transfers
all content with the user’s cookie value to his or
her new user ID.

Mount sidekiq dashboard to ‘admin/sidekiq’

Skip fetch_user_and_handle_guest_cookie for the
admin controllers.