Add subscription model.
Discussions will now be able to have subscriptions.
So will users. When a user creates or responds to
a discussion, they will become subscribed to them.

These will hook up to a notification manager that
will send out emails when others have responded to
discussions that people are participating in.

Users will also be able to subscribe to discussions
they have not responded to yet. As will guests by
providing their email address which will be saved
in the guest_email column.

Subscriptions can be disabled via email links via
the email_token which is generated when a sub is
created.