Add Notification and SubscriptionManager services.
NotificationService handles the delivery of email
notifications to users and admins when new
discussions are created by users, or when either
guests or users subscribe to discussions.

SubscriptionManagerService handles the subscribing
of users and guests (via an email form) to new and
existing discussions.

These services are accompanied by their respective
Sidekiq worker classes that send out their mailers
asynchronously. I don’t like that there are 3 new
workers, each dedicated to sending 1 email, but I
haven’t spent too much time investigating an 
alternative approach.

Add the initial emails for the notifications, as
well as the mailer methods.

In the discussion new and show pages, add fields 
for guest users to be able to subscribe by providing
their email address. Also add a (un)subscribe button
for regular users on the show page.

Use icomoon font glyphs and add some styling to the
back link on the discussions page.

Abstract the logic of displaying subscribe or 
unsubscribe in a discussion helper.

Feature tests for a range of notification as well
as subscription scenarios. Unit test the services.

Also update the controller tests so they stay green.